Components pass null through channels.
Js-CSP was preventing null from passing though a channel as it is interpreted as indicating a channel is closed.  You cannot put it on a channel, and when you take null from a channel it is interpreted as a signal that the channel is closed.  Null is a valid value to pass around in crkt.  It has semantic meaning in many situations, or to the API that is being coupled into a graph.  The work around involves passing a NULL_OBJECT through the channel, and transforming it on each end from/to null.